Refactor editor tree styles, add block actions
